Set 'value' as attribute, add a 'change' event
Elm can natively pass thing to the web-component as attributes, and
subscribe to events.

In order to make usage in a Elm application easy, this patch changes the
'value' property into an attribute, and dispatch an event 'change' with
the new value as a detail whenever the editor content changes.

If the 'value' attribute is changed, the editor value is updated
accordingly, but if the value is not different from the current one, the
editor is left untouched. This is to cope with the Elm way of handling
input values.